% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{attack.rate}
\alias{attack.rate}
\title{Compute attack rates}
\usage{
attack.rate(outcome, exposure, data, rowcol = "cols")
}
\arguments{
\item{outcome}{Variable for outcome values}

\item{exposure}{Variable to be tested as risk factor}

\item{data}{The dataset}

\item{rowcol}{Row or col percentage}
}
\value{
The result table
}
\description{
Compute attack rates for two by two tables
}
