% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC.R
\name{CC}
\alias{CC}
\title{CC function}
\usage{
CC(cases, exposure, exact = F, full = FALSE)
}
\arguments{
\item{cases}{cases vector}

\item{exposure}{vector}

\item{full}{display all}
}
\value{
The result
}
\description{
CC function
}
