% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetools.R
\name{chartodate}
\alias{chartodate}
\title{Transform character values into date}
\usage{
chartodate(x, format = "d/m/y", update = TRUE)
}
\arguments{
\item{x}{Character value to transform into date}

\item{format}{Optionnal format}

\item{update}{If TRUE the default, the data.frame is updtade with new values}
}
\value{
the changed data
}
\description{
Transform character values into date
}
\examples{
data(gastro)
chartodate(gastro$dob,"yy/m/d")
}
