% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetools.R
\name{clear}
\alias{clear}
\title{clear}
\usage{
clear(what, noask = FALSE)
}
\arguments{
\item{what}{vars, functions or all}

\item{noask}{to clear whithout confirmation. Useful when running from a script}
}
\description{
clear memory for beginner.
}
