% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetools.R
\name{count}
\alias{count}
\title{Title count}
\usage{
count(expr)
}
\arguments{
\item{expr}{A logical expression}
}
\value{
Number of rows macthing expr
}
\description{
count number of record / row corresponding to expr
}
\examples{
count(c(1, 2, 3, 1) == 1)

}
