% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataman.R
\name{describe}
\alias{describe}
\title{Title  List data.frame column}
\usage{
describe(df)
}
\arguments{
\item{df}{A data.frame}
}
\value{
Structure of data.frame column
}
\description{
Title  List data.frame column
}
\examples{
describe(test)
}
