% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetools.R
\name{dropvar}
\alias{dropvar}
\title{Title Drop a data.frame column}
\usage{
dropvar(varname)
}
\arguments{
\item{varname}{The name of the column to be dropped}
}
\value{
Message to confirm
}
\description{
Title Drop a data.frame column
}
\examples{
dropvar(dumepivar)
}
