% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epifield.R
\docType{package}
\name{epifield}
\alias{epifield}
\title{epifield}
\description{
Epi package is designed to help epidemiologiste by hiding some of the R complexity
  while providing advices
}
\details{
\emph{draft version}

epifield functions:
The epifield functions aim to reproduce some of EPI6 and STATA function
}
\author{
Gilles DESVE and al
}
