% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiorder.R
\name{epiorder}
\alias{epiorder}
\title{epiorder}
\usage{
epiorder(data, vars, values, labels = NULL)
}
\arguments{
\item{data}{data.frame}

\item{vars}{vector of colnames}

\item{values}{vector of ordered values}

\item{labels}{vector of optionals ordered labels}
}
\value{
data.frame
}
\description{
epiorder
}
\examples{
# TODO
}
\author{
Jean Pierre Decorps
}
