% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{epitable}
\alias{epitable}
\title{Cross tabulation ( 2by2 table).}
\usage{
epitable(out, exp, missing = FALSE, row = FALSE, col = FALSE,
  fisher = TRUE)
}
\arguments{
\item{out}{"Outcome" as numbers, factors or text}

\item{exp}{"Exposure" as numbers, factors or text. short syntax is available
see help(epifield)}

\item{missing}{Boolean if FALSE, missing are not included in the table.
A summary output of number of missing values is added at the end}

\item{row}{"Row percentages"}

\item{col}{"Col percentages"}

\item{fisher}{TRUE by default, display the fisher exact probability.
If table is larger than 2*2 then Fisher is not calculated}
}
\value{
An array containing  values of \code{...}   \code{}
}
\description{
\code{epitable} Display a cross tabulation of two variables optionnaly with
 row or col percentages. Chi Square with associated p.value are calculated.
 If table contain binary variable, then epiorder function is apply on the two variable
 to get a resulting table compatible with usual epidemiology interpretation.
 0/1 variables are transformed into Yes/No and Yes is displayed before No
 Exposed Cases appear on upper left part of the table.
}
\examples{
#' epitable(c(1,1,2,2,1),c(3,3,4,4,4))

}
\references{
Based on: \emph{Epi6} and \emph{Stata} functionnality,
available at \url{https://github.com/}.
}
\seealso{
\code{\link{freq}} for frequency distributions
}
\author{
Gilles Desve
}
