% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataman.R
\name{eval_expr}
\alias{eval_expr}
\title{Title  eval an expression looking at default dataset}
\usage{
eval_expr(expr, env)
}
\arguments{
\item{expr}{An expression to be evaluated}

\item{env}{An otionnal environnement to evaluate expr}
}
\value{
Result of expression evaluation
}
\description{
Title  eval an expression looking at default dataset
}
\examples{
eval_expr(5>4)
}
