% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetools.R
\docType{data}
\name{gastro}
\alias{gastro}
\title{#title Gastroenteritidis outbreak in pensylvania}
\format{
A data frame with 103 rows and 32 variables:
\describe{
  \item{id}{Sequential number}
  \item{dob}{Date of birth}
  \item{age}{Age (year)}
  \item{sex}{Sex Males / Females}
  \item{floor}{Floor of residence}
  \item{room}{Room number}
  \item{meal}{Meal in dinning room}
  \item{duration}{Duration of symptom}
  \item{dayonset}{Day of onset}
  \item{timeonset}{Time of onset}
  \item{diarrhoea}{Diarrhoea}
  \item{temperatur}{temperatur}
  \item{fever}{fever}
  \item{vomiting}{vomiting}
  \item{abdopain}{abdopain}
  \item{nausea}{nausea}
  \item{headache}{headache}
  \item{blended}{blended}
  \item{eggs}{eggs}
  \item{chicken}{chicken}
  \item{potatoes}{potatoes}
  \item{broccoli}{broccoli}
  \item{ham}{ham}
  \item{sandwich}{sandwich}
  \item{fruits}{fruits}
  \item{blendedfru}{blendedfru}
  \item{protein}{protein}
  \item{lab}{lab}
  \item{agegroup}{agegroup}
  \item{case}{case}
  \item{earlycase}{earlycase}
  \item{latecase}{latecase}

}
}
\source{
Alain Moren investigation & case study
}
\usage{
gastro
}
\description{
A dataset containing the data from an outbreak
}
\keyword{datasets}
