% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetools.R
\name{get_option}
\alias{get_option}
\title{get_option}
\usage{
get_option(op)
}
\arguments{
\item{op}{name of the option to retrieve}
}
\value{
option value
}
\description{
retrieve a package option
}
\examples{
get_option("option")


}
