% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetools.R
\name{getvar}
\alias{getvar}
\title{retrieve a data.frame column}
\usage{
getvar(what = NULL)
}
\arguments{
\item{what}{Name of the column}
}
\value{
The column
}
\description{
retrieve a data.frame column
}
\examples{
getvar()
}
