% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetools.R
\name{\%and\%}
\alias{\%and\%}
\title{Replacement for &}
\usage{
a \%and\% b
}
\arguments{
\item{a}{logical or condition}

\item{b}{logical or condition}
}
\description{
%and% can be used everywhere & is used.
To be tested
}
\details{
it's mainly for those whitout & on keyboard
}
\examples{
TRUE \%and\% TRUE

}
\references{
Based on: \emph{Epi6} and \emph{Stata} functionnality, available at \url{https://github.com/}.
}
\author{
Gilles Desve
}
