% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{histogram}
\alias{histogram}
\title{histogram}
\usage{
histogram(xvar, df = as.data.frame())
}
\arguments{
\item{xvar}{As numbers, factors or text.}

\item{df}{As data.frame}
}
\value{
An array containing  resulting graph
#' @examples
histogram(c(3,1,2,2,5))
}
\description{
\code{histogram} Display an histogram of frequency distribution.
#'
}
\references{
Based on: \emph{Epi6} and \emph{Stata} functionnality,
available at \url{https://github.com/}.
}
\seealso{
\code{\link{barchart}} for bar graph
}
\author{
Gilles Desve
}
