% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps.move.R
\name{move}
\alias{move}
\title{Move a column of a data.frame to another position}
\usage{
move(data, tomove, where = "last", target = NULL)
}
\arguments{
\item{data}{data.frame to be modified}

\item{tomove}{character - the column to be moved}

\item{where}{character - in c("first", "last", "before", "after")}

\item{target}{- character - column when where is "after" or "before"}
}
\value{
data.frame
}
\description{
Move a column of a data.frame to another position
}
\examples{
data(tira)
tira <- move(tira, "uniquekey", "after", "dateonset")
}
