% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{or}
\alias{or}
\title{calculate odds ratio and confidence intervales based on fisher exact probability}
\usage{
or(A, B = NULL, C = NULL, D = NULL)
}
\arguments{
\item{A}{A matrix or the first numer of a 2*2 table case exposed}

\item{B}{Second number of a 2*2 table  non case exposed}

\item{C}{Third number of a 2*2 table  control exposed}

\item{D}{Fourth number of a 2*2 table control non exposed}
}
\value{
a list with or, lci, uci
}
\description{
calculate odds ratio and confidence intervales based on fisher exact probability
}
\examples{
or(5,2,1,3)
}
\references{
Based on: \emph{Epi6} and \emph{Stata} functionnality,
available at \url{https://github.com/}.
}
\seealso{
\code{\link{freq}} for frequency distributions
}
\author{
Gilles Desve
}
