% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataman.R
\name{recode.if}
\alias{recode.if}
\title{Title recode a var if a condition if true}
\usage{
recode.if(xvar, condition, newvalue)
}
\arguments{
\item{xvar}{The vector or data.frame column to recode}

\item{condition}{Logical expression to select value to be recoded}

\item{newvalue}{The value to replace}
}
\value{
The recoded variable
}
\description{
Title recode a var if a condition if true
}
\examples{
recode.if(age, age > 100, 100)
}
