% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetools.R
\name{rename}
\alias{rename}
\title{Change the name of a data.frame column}
\usage{
rename(oldname, newname)
}
\arguments{
\item{oldname}{Name of the column/variable to rename}

\item{newname}{New name to apply}
}
\value{
Message to confirm the change
}
\description{
Change the name of a data.frame column
}
\examples{
df <- as.data.frame( c(One=1,Two=2) )
rename(Two,Last)
}
