% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetools.R
\name{right}
\alias{right}
\title{right}
\usage{
right(text, num_char)
}
\arguments{
\item{text}{Text to extract from}

\item{num_char}{Number of char to extract from rigth}
}
\value{
\code{num_char} extracted characters
}
\description{
Extract x rigth characters from a text
}
\examples{
\dontrun{
right("dummy_test",4)
}
}
