% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{savegraph}
\alias{savegraph}
\title{Title}
\usage{
savegraph(name, type = png)
}
\arguments{
\item{name}{Name of the file}

\item{type}{Type of saved image (png,jpeg,pdf,bmp,wmf)}
}
\value{
nothing
}
\description{
Title
}
\examples{
savegraph("test")
}
