% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetools.R
\name{select}
\alias{select}
\title{Title select part of a data.frame}
\usage{
select(expr)
}
\arguments{
\item{expr}{A logical condition applied to data.frame in order to select rows}
}
\value{
the selected vector
}
\description{
Title select part of a data.frame
}
