% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetools.R
\name{setdata}
\alias{setdata}
\title{setdata}
\usage{
setdata(df = NULL)
}
\arguments{
\item{df}{name of the dataframe to assign}
}
\value{
current df name
}
\description{
assign a data.frame as default dataframe
}
\examples{
df <-as.data.frame(c(1,2))
setdata(df)
rm(df)

}
