% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataman.R
\name{sumby}
\alias{sumby}
\title{Title}
\usage{
sumby(countvar, byvar, byvar2, byvar3)
}
\arguments{
\item{countvar}{Variable to be aggregated}

\item{byvar}{Indicator for wich sum will be calculated}

\item{byvar2}{Second Indicator for wich sum will be calculated}

\item{byvar3}{Third Indicator for wich sum will be calculated}
}
\value{
aggregated data
}
\description{
Title
}
\examples{
df <- data.frame("count"=c(4,5),"year"=c(2015,2016))
sumby(count,year)
}
