% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{sumstats}
\alias{sumstats}
\title{sumstats}
\usage{
sumstats(what, cond)
}
\arguments{
\item{what}{Variable to be analysed}

\item{cond}{Optionnal logical expression to filter data}
}
\value{
summary
}
\description{
Produce a summary
}
\examples{
sumstats(gastro$age)

}
