% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetools.R
\docType{data}
\name{test}
\alias{test}
\title{#title Test sample from outbreak}
\format{
A data frame with 10 rows and 5 variables:
\describe{
  \item{X}{Sequential number}
  \item{age}{Age of cases}
  \item{sex}{outcome variable}
  \item{tira}{outcome variable}
  \item{beer}{outcome variable}

}
}
\source{
Epiet case study
}
\usage{
test
}
\description{
A dataset containing somedata for testing purpose
}
\keyword{datasets}
