% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetools.R
\docType{data}
\name{tira}
\alias{tira}
\title{#title Tiramisu outbreak}
\format{A data frame with 291 rows and 22 variables:
\describe{
  \item{X}{Sequential number}
  \item{age}{Age of cases}
  \item{ill}{outcome variable}
  \item{dateonset}{outcome variable}
  \item{sex}{outcome variable}
  \item{tira}{outcome variable}
  \item{tportion}{outcome variable}
  \item{wmousse}{outcome variable}
  \item{dmousse}{outcome variable}
  \item{mousse}{outcome variable}
  \item{mportion}{outcome variable}
  \item{beer}{outcome variable}
  \item{uniquekey}{outcome variable}
  \item{redjelly}{outcome variable}
  \item{fruitsalad}{outcome variable}
  \item{tomato}{outcome variable}
  \item{mince}{outcome variable}
  \item{salmon}{outcome variable}
  \item{horseradish}{outcome variable}
  \item{chickenwin}{outcome variable}
  \item{roastbeef}{outcome variable}
  \item{pork}{outcome variable}

}}
\source{
Epiet case study
}
\usage{
tira
}
\description{
A dataset containing the data from an outbreak
}
\keyword{datasets}
