% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetools.R
\name{total}
\alias{total}
\title{count number of record / row}
\usage{
total(expr)
}
\arguments{
\item{expr}{A logical expression}
}
\value{
Number of rows macthing expr
}
\description{
Toyal return the number of row sastifying a condition.
}
\examples{
total(c(1, 2, 3, 1) == 1)

}
