% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{unfactor}
\alias{unfactor}
\title{Title}
\usage{
unfactor(xvar, levels, update = TRUE)
}
\arguments{
\item{xvar}{The variable to unfactor}

\item{levels}{The desired levels as numeric values given in the order where
label appear}

\item{update}{if TRUE, the default, the data.frame is automaticaly updated
with the new numeric variable}
}
\value{
The unfactored variable
}
\description{
Title
}
\examples{
tvar<- c("Yes","No")
tvar <- factor(tvar)
unfactor(tvar)
}
