% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetools.R
\name{use}
\alias{use}
\title{use}
\usage{
use(filename = "", label = NULL)
}
\arguments{
\item{filename}{Name of file to be read. Type is defined by extension}

\item{label}{Label to be added as attribute to dataframe}
}
\description{
read a data.frame.
 The function try to identify the file structure in order to call the appropriate specific
 command
}
\examples{
fil <- tempfile(fileext = ".data")
cat("TITLE extra line", "2 3 5 7", "", "11 13 17", file = fil,
sep = "\\n")
readLines(fil, n = -1)
unlink(fil) # tidy up

}
