% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataman.R
\name{view}
\alias{view}
\title{List head of data.frame allowing to select column to view}
\usage{
view(..., nline = 10)
}
\arguments{
\item{...}{variable name to include in list}

\item{nline}{Number of rows to show (default to n) If negative then number of row from the end}
}
\value{
the resulting list
}
\description{
List head of data.frame allowing to select column to view
}
\examples{
\dontrun{
view(age,case)
view(age,case, nline=5)
view(age,case, nline= -10)
}
}
