% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bargraph}
\alias{bargraph}
\title{bargraph}
\usage{
bargraph(xvar, title)
}
\arguments{
\item{xvar}{As numbers, factors or text.}

\item{title}{As character : main title}
}
\value{
An array containing  resulting graph
#' @examples
bargraph(c(3,1,2,2,5))
}
\description{
\code{bargraph} Display a bar graph of frequency distribution.
#'
}
\references{
Based on: \emph{Epi6} and \emph{Stata} functionnality,
available at \url{https://github.com/}.
}
\seealso{
\code{\link{histogram}} for bar graph
}
\author{
Gilles Desve
}
