% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmissing.R
\name{checkMissing}
\alias{checkMissing}
\title{checkMissing}
\usage{
checkMissing(data, vars = NULL, sort = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{vars}{NULL or a vector of colnames}

\item{sort}{boolean If TRUE the output table is sorted}
}
\value{
list of 2 tables
}
\description{
checkMissing
}
\examples{
# still nothing
}
\author{
Jean Pierre Decorps
}
