% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetools.R
\name{clear}
\alias{clear}
\title{Remove variables, dataset or functions from memory}
\usage{
clear(what, noask = FALSE)
}
\arguments{
\item{what}{Keyword (vars, functions, all) or pattern}

\item{noask}{to clear whithout confirmation. Useful when running from a script}
}
\description{
Clear can be used to remove objects from memory (variables, data.frame, functions).
Clear is easier than \code{\link{rm}} and is more secure because, by default, it ask for confirmation.
Objects to remove can be specified as is or by their name ("character").
It's possible to erase all vars, all functions using keywords : "vars" or "functions"
"all" keyword will allows total cleaning.
}
\details{
When keyword or pattern are used and there is more than one object to clear, a confirmation will be issued.
Except if noask parameters is set to true
If there only one object matching the exactly the \code{what} parameter, this object is removed whithout confirmation
}
\examples{
tmp <- 5
temp <- 5
clear(t)

}
\references{
Based on: \emph{Epi6} and \emph{Stata} functionnality, available at \url{https://github.com/}.
}
\seealso{
\code{\link{rm}}
}
\author{
Gilles Desve
}
