% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetools.R
\name{count}
\alias{count}
\title{count number of record / row}
\usage{
count(expr)
}
\arguments{
\item{expr}{A logical expression ora data.frame.I f a data.frame is given,
count() return the number of rows. If a logical expression is given, total() return
the number of records satisfaying the condition}
}
\value{
Number of rows macthing expr
}
\description{
Count return the number of row sastifying a condition.
}
\examples{
count(c(1, 2, 3, 1) == 1)

}
