% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{epiorder}
\alias{epiorder}
\title{Reorder data for epi table ( 2by2 table).}
\usage{
epiorder(var, mode = "Yesno", custom = NULL)
}
\arguments{
\item{var}{Variable to reorder (will be converted as factor).}

\item{mode}{"Yesno"   "Yesno"  "10" "+-"}

\item{custom}{Custom labels}
}
\value{
A vector reordered
}
\description{
\code{epiorder} Rearrange order of factor variable to produce classical epi table
 1/0  Yes/No  +/-
}
\examples{
\dontrun{
epiorder(c(0,1,0,1,1))
}


}
\references{
Based on: \emph{Epi6} and \emph{Stata} functionnality,
available at \url{https://github.com/}.
}
\seealso{
\code{\link{epitable}} for cross tabulation
}
\author{
Gilles Desve
}
