% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{epiorder}
\alias{epiorder}
\title{Reorder data for epi table ( 2by2 table).}
\usage{
epiorder(var, mode = "yesno", levels = NULL, update = TRUE,
  reverse = FALSE)
}
\arguments{
\item{var}{Variable to reorder (will be converted as factor).}

\item{mode}{Label plan for the new factor variable
"yesno" for Yes, No
"10"    for 1, 0
"+-"    for +,-
"truefalse"  for TRUE, FALSE
or any set of two labels on the form c("A","B")}

\item{levels}{Custom set of levels as vector : c(1,0)
This levels will replaced by the lables then levels and labels should have the same
length and the same order}

\item{update}{if TRUE (the default) Then the original dataset is updated with new value
The recoded column is replaced by the returned value
With this option, there is no need to reassign the retruned value,
but original data are lost.}

\item{reverse}{if TRUE labels are reordered to better fit epidemiological tables
with 1 before 0 , Yes before No etc...
Other type of label are not changed, existing factor are not changed}
}
\value{
A vector reordered Can be nested as in \code{freq(epioredr(case))}
}
\description{
\code{epiorder} Rearrange order of factor variable to produce classical epi table
 1/0  Yes/No  +/-
}
\examples{
\dontrun{
epiorder(c(0,1,0,1,1))
}


}
\references{
Based on: \emph{Epi6} and \emph{Stata} functionnality,
available at \url{https://github.com/}.
}
\seealso{
\code{\link{epitable}} for cross tabulation
}
\author{
Gilles Desve
}
