% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{epitable}
\alias{epitable}
\title{Cross tabulation ( 2by2 table).}
\usage{
epitable(exp, out, row = FALSE, col = FALSE)
}
\arguments{
\item{exp}{"Exposure" as numbers, factors or text.}

\item{out}{"Outcome" as numbers, factors or text}
}
\value{
An array containing  values of \code{...}   \code{}
}
\description{
\code{epitable} Display a cross tabulation of two binary variables optionnaly with
 row or col percentages.
}
\examples{
\dontrun{
epitable(c(3,1,2,2,5))
}


}
\references{
Based on: \emph{Epi6} and \emph{Stata} functionnality,
available at \url{https://github.com/}.
}
\seealso{
\code{\link{freq}} for frequency distributions
}
\author{
Gilles Desve
}
