% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{histogram}
\alias{histogram}
\title{histogram}
\usage{
histogram(xvar, title, ylab = "count", xlab, by = "days", width = 1,
  color = "#000099")
}
\arguments{
\item{xvar}{As numbers, factors or text.}

\item{title}{As character : main title}

\item{ylab}{Y axis label}

\item{xlab}{X axis label}

\item{by}{For date values by parameters allows to plot histogram "by" days,months or years}

\item{width}{numeric, define width of each bar using the unit of the variable
width = 5 to have histogram of age by 5 years}

\item{color}{Colors can be specified as color eg "red" "blue" or as a hexadecimal RGB triplet,
such as "#0066CC". The first two digits are the level of red, the next two green, and the last two blue.
The value for each ranges from 00 to FF in hexadecimal (base-16) notation,
which is equivalent to 0 and 255 in base-10.
For example, “#FFFFFF” is white and “#990000” is a deep red.}
}
\value{
An array containing  resulting graph
#' @examples
histogram(c(3,1,2,2,5))
}
\description{
\code{histogram} Display an histogram of frequency distribution.
#'
}
\references{
Based on: \emph{Epi6} and \emph{Stata} functionnality,
available at \url{https://github.com/}.
}
\seealso{
\code{\link{bargraph}} for bar graph
}
\author{
Gilles Desve
}
