% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{incrates}
\alias{incrates}
\title{Calculate incidence rates with CI}
\usage{
incrates(count, denom, data, per = 1e+05, conflvl = 0.95)
}
\arguments{
\item{count}{variable to be used as countfor calculating the incidence}

\item{denom}{denominator to use for incidence rates}

\item{data}{Optionnal data.frame containing the data}

\item{per}{integer to use as unit per}

\item{conflvl}{The confidence level, usualy 0.95}
}
\value{
a table of result
}
\description{
Calculate incidence rates with CI
}
\examples{
dat <- data.frame("lab"=c("A","B","C"),"nb"=c(5,10,20),"den"=c(100,120,150))
incrates(dat$lab,dat$nb,dat$den)
}
