% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{incrates}
\alias{incrates}
\title{Title Calculate incidence rates with CI}
\usage{
incrates(df, conflvl = 0.95)
}
\arguments{
\item{df}{A 2 column data.frame}

\item{conflvl}{The confidence level, usualy 0.95}
}
\value{
a table of result
}
\description{
Title Calculate incidence rates with CI
}
\examples{
dat <- cbind(c(5,10,20),c(100,120,150))
incrates(dat)
}
