% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataman.R
\name{recode.if}
\alias{recode.if}
\title{Title recode a var if a condition if true}
\usage{
recode.if(var, value, condition)
}
\arguments{
\item{var}{The vector or data.frame column to recode}

\item{value}{The value to assign}

\item{condition}{If the condition is verified}
}
\value{
The recode variable
}
\description{
Title recode a var if a condition if true
}
\examples{
recode.if(age, "young", age < 18)
}
