% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataman.R
\name{recode.value}
\alias{recode.value}
\title{Title recode a var if a condition if true}
\usage{
recode.value(xvar, oldvalue, newvalue)
}
\arguments{
\item{xvar}{The vector or data.frame column to recode}

\item{oldvalue}{The value to replace}

\item{newvalue}{The value to assign}
}
\value{
The recode variable
}
\description{
Title recode a var if a condition if true
}
\examples{
recode.if(age, "Unknow", NA)
}
