% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataman.R
\name{recode.value}
\alias{recode.value}
\title{Title recode a var if a condition if true}
\usage{
recode.value(var, oldvalue, newvalue)
}
\arguments{
\item{var}{The vector or data.frame column to recode}

\item{oldvalue}{The value to assign}

\item{newvalue}{The value to replace}
}
\value{
The recode variable
}
\description{
Title recode a var if a condition if true
}
\examples{
recode.value(case, "Oui", 1)
}
