% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps.rename.R
\name{rename}
\alias{rename}
\title{Title}
\usage{
rename(data, colname, newname)
}
\arguments{
\item{data}{data.frame}

\item{colname}{chr - name of the column}

\item{newname}{chr - new name of the column}
}
\value{
data.frame
}
\description{
Title
}
\examples{
#
}
