% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetools.R
\name{set_option}
\alias{set_option}
\title{set_option}
\usage{
set_option(op, value)
}
\arguments{
\item{op}{name of the option to assign}

\item{value}{The value to be assigned to option}
}
\value{
The previous option value before new assignment
}
\description{
assign a package option
}
\examples{
set_option("option",1)


}
